+++ CHANGELOG.md
@@ -27,8 +27,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Laravel storage adapters
 
-With this version empty responses won't be cached anymore by default.  
-You can enable caching for empty responses in the config by setting `allow_empty` to `true`.
+You can disable caching for empty responses in the config now by setting `allow_empty` to `false`.
 
 Check out the [docs](https://github.com/hamburgscleanest/guzzle-advanced-throttle#laravel-drivers) for more information on how to set it.