+++ CHANGELOG.md
@@ -23,6 +23,14 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ----------
 
+## 3.2.0
+
+### Added
+- Compatibility with Laravel / Illuminate 5.8
+- Upgraded PHPUnit to version 8
+
+----------
+
 ## 3.1.0
 
 ### Fixed